Entity package'ın,package ismi küçük harflerle değiştirildi ve gerekli getter ve setterlar eklendi.
