--- src/main/java/com/example/ciftcikredidegerlemesistemi/Entity/BaseEntity.java
@@ -1,8 +1,0 @@
-package com.example.ciftcikredidegerlemesistemi.Entity;
-
-import java.util.Date;
-
-public class BaseEntity {
-    private Date kayitTarihi;
-    private Date guncellemeTarihi;
-}
