+++ src/main/java/com/example/ciftcikredidegerlemesistemi/entity/HayvanIrk.java
@@ -1,14 +1,21 @@
-package com.example.ciftcikredidegerlemesistemi.Entity;
+package com.example.ciftcikredidegerlemesistemi.entity;
+import com.example.ciftcikredidegerlemesistemi.enums.HayvanTuru;
 import jakarta.persistence.*;
-import lombok.Data;
+import lombok.*;
+
 import java.util.List;
-@Data
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
 @Entity
+@Table(name = "HayvanIrk")
 public class HayvanIrk extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
     private String ad;
+    @Enumerated(EnumType.STRING)
     private HayvanTuru hayvanTuru;
     @OneToMany(mappedBy = "hayvanIrk", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<HayvansalVarlikDeger> hayvansalVarlikDegerler;
