+++ src/main/java/com/example/ciftcikredidegerlemesistemi/entity/HayvansalVarlikDeger.java
@@ -1,16 +1,24 @@
-package com.example.ciftcikredidegerlemesistemi.Entity;
+package com.example.ciftcikredidegerlemesistemi.entity;
+import com.example.ciftcikredidegerlemesistemi.enums.Cinsiyet;
 import jakarta.persistence.*;
-import lombok.Data;
+import lombok.*;
+
 import java.math.BigDecimal;
-@Data
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
 @Entity
+@Table(name = "HayvansalVarlikDeger")
 public class HayvansalVarlikDeger extends BaseEntity{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
+    @Enumerated(EnumType.STRING)
     private Cinsiyet cinsiyet;
     private BigDecimal deger;
     @ManyToOne
     @JoinColumn(name = "hayvan_irk_id")
     private HayvanIrk hayvanIrk;
+
 }
